+++ packages/terra-button-group/CHANGELOG.md
@@ -4,6 +4,9 @@
 * Changed
   * Replaced `terra-button` with `IconButton`
 
+* Added
+  * Added keyboard navigation feature and fixes to convey selection states.
+
 ## 3.60.2 - (February 15, 2022)
 
 * Changed